Add HttpTaskExtensions class for parsing JSON responses
Update ToJsonContent method to return StringContent

Throw exception if wallet is not found in WalletRepository
